Split Local Login and LDAPLogin into two components
